remove proxydetectionprovider and move detection method to networkmetadata provider.
